+++ back-end/test-rules/tools/clearAndPrime.cjs
@@ -1,10 +1,9 @@
 /*
 * back-end/test-rules/tools/clearAndPrime.cjs
 *
-* Using '@firebase/testing' library, clear the existing data and replace with given.
+* Using 'firebase-admin' library, clear the existing data and replace with given.
 *
-* NOTE: Once this is possible using ES modules, we can consider placing it as part of 'firebase-jest-testing'.
-*     For now, it remains within the sample.
+* NOTE: Once Jest 'globalSetup' can use ES modules, we should make this one.
 */
 //import { strict as assert } from 'assert'
 const assert = require('assert').strict;
@@ -12,30 +11,23 @@ const assert = require('assert').strict;
 const PRIME_ROUND = !global.afterAll;   // are we called from 'globalSetup'
 assert(PRIME_ROUND);
 
-// import { clearAll, dbAdmin } from 'firebase-jest-testing/firestoreTesting/admin'   // something like so, later..
-const firestoreTestingAdmin = require('../hack-cjs/firestoreTestingAdmin.cjs');
+// import { clearAll, prime } from 'firebase-jest-testing/fireStore'
+const cjsTools = require('firebase-jest-testing/cjs'); const { clearFirestoreData, prime } = cjsTools;
+
+// We just know it (from 'firebase-jest-testing')   #later: once we're in ESM, let's make an API that exposes it?
+//
+const projectId = "rules-test";
 
 /*
 * Clear a database and prime it with data
 *
 * Note: It is enough to call this once, per Jest test run (via 'globalSetup').
 */
-async function clearAndPrime(projectId, docs) {    // (string, { <docPath>: { <field>: <value> } }) => Promise of ()
-  assert(PRIME_ROUND);
+async function clearAndPrime(docs) {    // ({ <docPath>: { <field>: <value> } }) => Promise of ()
+  await clearFirestoreData({projectId});
+  await prime(projectId, docs);
 
-  const fst = firestoreTestingAdmin(projectId);
-  fst.clearAll();   // clear any earlier data
-
-  const {db} = fst;
-
-  const batch = db.batch();
-
-  for (const [docPath,value] of Object.entries(docs)) {
-    batch.set( db.doc(docPath), value );
-  }
-  await batch.commit();
-
-  /*await*/ const ignore = db.app.delete();    // clean the app; data remains (free running Promise)
+  console.debug("Cleared and primed!");
 }
 
 module.exports = clearAndPrime;